captchamb/requirements.txt
fastapi
uvicorn
pillow
mb-capcha-ocr==0.1.4
